fix: close native mpv platform-view teardown races, dedupe mpv backends
Addresses code review findings on the native mpv Apple playback PR:

Correctness (crash risk):
- iOS/macOS MpvPlayerCore.dispose() now waits for mpv_terminate_destroy
  to finish before resolving, matching tvOS's completion-based fix.
  Without this, a platform view could be unmounted while mpv's render
  thread still held an unretained pointer into its layer.
- PlaybackOrchestrator releases a failed PlatformViewProvider adapter's
  native view before falling through to the next backend, closing the
  same race on the automatic mpv -> AvKit fallback path.
- Multiview tile removal now awaits native teardown before unmounting
  the tile's platform view, instead of firing dispose() and unmounting
  in the same frame.

Docs/test accuracy:
- Fixed macOS/iOS feasibility docs and backendOrder to describe the
  fallback chains actually registered in buildPlaybackOrchestrator(),
  removed dangling references to a nonexistent MPV_MIGRATION_STATUS.md.

Cleanup:
- Extracted MpvNativeBackendBase/MpvNativeEvent to share the
  load/dispose/event-reduction FSM between Mac/AppleMpvNativeBackend
  instead of duplicating it, consolidated the two UnavailableException
  classes, shared a NativeVideoSurface widget between the player and
  Multiview screens, removed dead mpv log-message handling, throttled
  the native snapshot rebuild to skip the track-list walk on plain
  position ticks, and fixed em dashes introduced by this PR.
