+++ flutter_client/lib/main.dart
@@ -29,11 +29,13 @@ Future<void> main() async {
   final systemUiPolicy = SystemUiPolicy();
   await systemUiPolicy.applyBrowsing();
   // MediaKit (libmpv) is only bundled for Linux/Windows now -- Apple
-  // platforms use native mpv (MPVKit) as primary with AVKit as fallback.
+  // platforms use native mpv (MPVKit) as primary, with AVKit as an
+  // automatic fallback on iOS/tvOS (macOS has none, see
+  // buildPlaybackOrchestrator in lib/navigation/app_router.dart).
   // media_kit_libs_macos_video/ios_video were removed because they vendored
   // a second, independently-versioned ffmpeg/libmpv build that collided at
   // link time with MPVKit's, corrupting native mpv's own library-version
-  // check. See MPV_MIGRATION_STATUS.md.
+  // check.
   if (Platform.isLinux || Platform.isWindows) {
     MediaKit.ensureInitialized();
   }
