+++ flutter_client/lib/playback/player_adapter.dart
@@ -34,6 +34,15 @@ abstract class PlatformViewProvider {
 
   /// Creation-time arguments passed to the platform view factory, if any.
   Map<String, dynamic>? get platformViewCreationParams;
+
+  /// Tears down the native resources backing this platform view (e.g. a
+  /// native player core holding an unretained pointer into the view's
+  /// layer) without closing this adapter's [PlayerAdapter.onState]/
+  /// [PlayerAdapter.onError] streams, so the adapter is still safe to
+  /// [PlayerAdapter.load] again later. Must not resolve until native
+  /// teardown has actually finished releasing that pointer -- callers rely
+  /// on this to unmount the platform view safely afterward.
+  Future<void> releaseNativeView();
 }
 
 /// A [PlayerAdapter] that Multiview can drive: one concurrently playable
@@ -347,6 +356,18 @@ class PlaybackException implements Exception {
 
   @override
   String toString() => 'PlaybackException($backend, $code): $message';
+}
+
+/// A native mpv `PlatformViewProvider` backend's native core failed to
+/// initialize (e.g. `mpv_create`/`mpv_initialize` failed on the Swift side).
+/// Shared between `MacMpvNativeBackend` and `AppleMpvNativeBackend`.
+class NativeMpvUnavailableException extends PlaybackException {
+  NativeMpvUnavailableException(
+    String message, {
+    required super.backend,
+  }) : super(message: message, code: unavailableCode, recoverable: true);
+
+  static const String unavailableCode = 'backend_unavailable';
 }
 
 class _Unchanged {
