+++ flutter_client/tvos/Runner/MpvPlayer/MpvPlayerCore.swift
@@ -240,12 +240,18 @@ final class MpvPlayerCore {
       emit(kind: "START_FILE", extra: [:])
     case MPV_EVENT_FILE_LOADED:
       readyEmitted = true
-      emit(kind: "FILE_LOADED", extra: snapshot())
+      emit(kind: "FILE_LOADED", extra: snapshot(includeTracks: true))
     case MPV_EVENT_PLAYBACK_RESTART:
-      emit(kind: "PLAYBACK_RESTART", extra: snapshot())
+      emit(kind: "PLAYBACK_RESTART", extra: snapshot(includeTracks: true))
     case MPV_EVENT_PROPERTY_CHANGE:
       if readyEmitted {
-        emit(kind: "PLAYBACK_RESTART", extra: snapshot())
+        // Most property-change events are `time-pos` ticks (essentially
+        // every frame during playback); only re-walk the track list when
+        // the property that actually changed is track-related, so a
+        // position tick doesn't pay for an mpv_get_property(track-list)
+        // NODE walk + Dart-side re-parse dozens of times a second.
+        let includeTracks = Self.trackRelatedPropertyIndices.contains(event.reply_userdata)
+        emit(kind: "PLAYBACK_RESTART", extra: snapshot(includeTracks: includeTracks))
       }
     case MPV_EVENT_END_FILE:
       if let data = event.data {
@@ -264,7 +270,12 @@ final class MpvPlayerCore {
     }
   }
 
-  private func snapshot() -> [String: Any] {
+  /// Indices passed to `mpv_observe_property` in `attach(to:)` for
+  /// properties whose change should trigger a `track-list` re-walk in
+  /// `snapshot(includeTracks:)` -- `aid`, `sid`, `track-list`.
+  private static let trackRelatedPropertyIndices: Set<UInt64> = [6, 7, 8]
+
+  private func snapshot(includeTracks: Bool) -> [String: Any] {
     guard let handle = mpv else { return [:] }
     var result: [String: Any] = [:]
 
@@ -300,11 +311,13 @@ final class MpvPlayerCore {
       result["sid"] = sid
     }
 
-    let tracks = trackList(handle)
-    result["audioTracks"] = tracks.filter { $0["type"] as? String == "audio" }
-      .map { ["id": $0["id"] as Any, "label": $0["label"] as Any, "language": $0["language"] as Any] }
-    result["subtitleTracks"] = tracks.filter { $0["type"] as? String == "sub" }
-      .map { ["id": $0["id"] as Any, "label": $0["label"] as Any, "language": $0["language"] as Any] }
+    if includeTracks {
+      let tracks = trackList(handle)
+      result["audioTracks"] = tracks.filter { $0["type"] as? String == "audio" }
+        .map { ["id": $0["id"] as Any, "label": $0["label"] as Any, "language": $0["language"] as Any] }
+      result["subtitleTracks"] = tracks.filter { $0["type"] as? String == "sub" }
+        .map { ["id": $0["id"] as Any, "label": $0["label"] as Any, "language": $0["language"] as Any] }
+    }
 
     return result
   }
